test(cli): extract progression read play tests
Move the traditions and progress-dashboard read tests into a focused play suite. The new file owns the readTraditionsView/readProgressDashboard fake responses and progression payload builders.

Keep the monolith responsible for tradition-review notification, notification queue, and compact-priority projection fixtures. Wire the focused suite into test:cli:play so the extracted coverage remains canonical.

Proof: git diff --check; bun run --cwd packages/cli test -- game.play.progression-read.test.ts; bun run --cwd packages/cli test -- game.play.test.ts -t "tradition|progress-dashboard|compact priorities|notification queue"; bun run check:cli; bun run test:cli:play. This is test ownership modularization only; it changes no runtime behavior, direct-control behavior, live-game flow, gameplay send, play-thread workflow, Studio/oRPC transport, or relationship/city-state/suzerain evidence policy.
